refactor(webapp): pay down boy-scout debt in backend-local.ts
Move layer-up imports out of the fs mount backend: approval-card HTML
and popup/direct acquisition stay in fs/, cone tool-UI flow moves to
shell with injection through MountCommands. Clears layer-back-edge and
record-string-unknown baselines for this file.
