+++ packages/dev-tools/tools/record-string-unknown-baseline.json
@@ -6,7 +6,6 @@
   "packages/webapp/providers/github-copilot.ts": 2,
   "packages/webapp/providers/openai-codex.ts": 2,
   "packages/webapp/src/cdp/har-recorder.ts": 10,
-  "packages/webapp/src/fs/mount/backend-local.ts": 1,
   "packages/webapp/src/kernel/realm/py-realm-shared.ts": 1,
   "packages/webapp/src/providers/built-in/bedrock-camp.ts": 31,
   "packages/webapp/src/scoops/tray-leader/cdp-router.ts": 5,
