+++ packages/webapp/src/shell/supplemental-commands/picker-approval.ts
@@ -12,57 +12,18 @@
  *
  * Each kind shares the same card markup (title + Approve/Deny buttons +
  * `data-picker=<kind>` attribute carrying the filters); per-kind text
- * comes from {@link PICKER_KIND_TEXT}.
+ * comes from `fs/picker-approval-card.ts`.
  */
 
-import { escapeHtml } from '@slicc/webcomponents/internal/html';
+import { buildApprovalCardHtml } from '../../fs/picker-approval-card.js';
 import { showToolUI, type ToolExecutionContext, toolUIRegistry } from '../tool-ui.js';
-import type { PickerKind } from './picker-popup.js';
+
+export { buildApprovalCardHtml };
 
 /** Two minutes — enough for a slow user, short enough to fail loud. */
 const APPROVAL_TIMEOUT_MS = 120_000;
 
 const APPROVAL_TIMEOUT_SENTINEL = Symbol('picker-approval-timeout');
-
-interface PickerKindText {
-  title: string;
-  approve: string;
-}
-
-const PICKER_KIND_TEXT: Record<PickerKind, PickerKindText> = {
-  directory: { title: 'Mount local directory', approve: 'Select directory' },
-  'usb-device': { title: 'Connect USB device', approve: 'Select USB device' },
-  'serial-port': { title: 'Connect serial port', approve: 'Select serial port' },
-  'hid-device': { title: 'Connect HID device', approve: 'Select HID device' },
-};
-
-/** Build the standard approval-card HTML for a picker kind + filters. */
-export function buildApprovalCardHtml(
-  kind: PickerKind,
-  filters: unknown[] = [],
-  targetPath?: string
-): string {
-  const text = PICKER_KIND_TEXT[kind];
-  const dataAttr = filters.length
-    ? ` data-action-data='${JSON.stringify({ filters }).replace(/'/g, '&apos;')}'`
-    : '';
-  const metaHtml =
-    kind === 'directory' && targetPath
-      ? `<div class="sprinkle-action-card__meta">Target: ${escapeHtml(targetPath)}</div>`
-      : '';
-  return `
-    <div class="sprinkle-action-card">
-      <div class="sprinkle-action-card__header">
-        <div class="sprinkle-action-card__title-group">${text.title}${metaHtml}</div>
-        <span class="sprinkle-badge sprinkle-badge--notice">approval</span>
-      </div>
-      <div class="sprinkle-action-card__actions">
-        <button class="sprinkle-btn sprinkle-btn--secondary" data-action="deny">Deny</button>
-        <button class="sprinkle-btn sprinkle-btn--primary" data-action="approve" data-picker="${kind}"${dataAttr}>${text.approve}</button>
-      </div>
-    </div>
-  `;
-}
 
 /** Serializable device descriptor returned by browser picker APIs. */
 export interface DevicePickerInfo {
